Rename 'segment' package to 'segments' to reflect folder name
