+++ pkg/storage/segments/common.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2018 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package segment
+package segments
 
 import (
 	"github.com/zeebo/errs"
